requirements.txt
absl-py==0.1.10
adal==1.2.0
antlr4-python3-runtime==4.7.1
applicationinsights==0.11.7
appnope==0.1.0
argcomplete==1.9.4
asn1crypto==0.24.0
astor==0.7.1
attrs==17.4.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.5
bleach==1.5.0
boto==2.48.0
boto3==1.5.35
botocore==1.8.49
bz2file==0.98
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.2.1
dill==0.2.8.2
diskcache==3.0.6
docker==3.6.0
docker-pycreds==0.4.0
docutils==0.14
entrypoints==0.2.3
funcy==1.11
future==0.16.0
gast==0.2.2
gensim==3.6.0
gitdb2==2.0.5
GitPython==2.1.11
googleapis-common-protos==1.5.3
grpcio==1.12.1
h5py==2.7.1
html5lib==0.9999999
humanfriendly==4.17
hyperopt==0.1
idna==2.6
imageio==2.4.1
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
isodate==0.6.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.5
jsonpickle==1.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
knack==0.5.1
liac-arff==2.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
mlxtend==0.15.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
msrest==0.6.2
msrestazure==0.5.1
murmurhash==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.5.1
networkx==1.11
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numexpr==2.6.8
numpy==1.15.3
numpy-indexed==0.3.5
oauthlib==2.1.0
overrides==1.9
pandas==0.22.0
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.1.1
pathspec==0.5.9
patsy==0.5.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.3.0
pkg-resources==0.0.0
plac==0.9.6
pluggy==0.6.0
portalocker==1.2.1
preshed==2.0.1
progressbar33==2.4
prompt-toolkit==1.0.15
protobuf==3.7.0
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
pyLDAvis==2.1.2
pymongo==3.6.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
pytest==3.4.1
python-dateutil==2.7.5
pytz==2018.3
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
regex==2018.1.10
requests==2.21.0
requests-oauthlib==1.0.0
ruamel.yaml==0.15.51
s3transfer==0.1.13
scikit-learn==0.19.1
scikit-optimize==0.5.2
scipy==1.0.0
seaborn==0.9.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
smart-open==1.5.6
smmap2==2.0.5
spacy==2.0.16
statsmodels==0.9.0
tabulate==0.8.2
tensorboard==1.11.0
tensorflow-gpu==1.11.0
tensorflow-hub==0.3.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.12.0
toolz==0.9.0
tornado==4.5.3
tqdm==4.28.1
traitlets==4.3.2
ujson==1.35
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.14.1
widgetsnbextension==3.1.4
wrapt==1.10.11
